requirements.txt
Flask
Flask-RESTx
Flask-SQLAlchemy
psycopg2-binary